Apr. 25

+++ main/student_registration/student_registration.php
@@ -4,6 +4,7 @@
         $studFname = htmlspecialchars($_POST['fname']);
         $studMname = htmlspecialchars($_POST['midname']);
         $studLname = htmlspecialchars($_POST['lname']);
+        $studLRN = htmlspecialchars($_POST['lrn']);
         $studSection = htmlspecialchars(strtolower($_POST['section']));
         $studStrand = htmlspecialchars($_POST['strand']);
         $studAdviser = htmlspecialchars($_POST['adviser']);
@@ -11,46 +12,51 @@
         $studAge = htmlspecialchars($_POST['age']);
         $studGrdLvl = htmlspecialchars($_POST['grd_lvl']);
         $studToken = uniqid("",true);
-
-        
-        $uploadDirectory = "../../assets/profile/";
-        if(isset($_FILES['stud-pfp'])){
-            $studPfp = $_FILES['stud-pfp'];
-        // Get the temporary file name
-        $tmpFileName = $studPfp["tmp_name"];
-        $img_ex = pathinfo($studPfp['name'], PATHINFO_EXTENSION);
-        $img_ex_to_lc = strtolower($img_ex);
-
-        $allowed_exs = array('jpg', 'jpeg', 'png');
-        if (in_array($img_ex_to_lc, $allowed_exs)) {
-            $new_img_name = uniqid($studLname, true) . '.' . $img_ex_to_lc;
-            $img_upload_path = $uploadDirectory . $new_img_name;
-
-            if (move_uploaded_file($tmpFileName, $img_upload_path)) {
-                echo "File uploaded successfully. Stored at: $new_img_name";
-            } else {
-                echo "Error uploading file.";
+        if(userExist($conn, $studFname, $studMname, $studLname)) {
+            echo "Student already exists";
+        } else {
+            $uploadDirectory = "../../assets/profile/";
+            if(isset($_FILES['stud-pfp'])){
+                $studPfp = $_FILES['stud-pfp'];
+            // Get the temporary file name
+            $tmpFileName = $studPfp["tmp_name"];
+            $img_ex = pathinfo($studPfp['name'], PATHINFO_EXTENSION);
+            $img_ex_to_lc = strtolower($img_ex);
+    
+            $allowed_exs = array('jpg', 'jpeg', 'png');
+            if (in_array($img_ex_to_lc, $allowed_exs)) {
+                $new_img_name = uniqid($studLname, true) . '.' . $img_ex_to_lc;
+                $img_upload_path = $uploadDirectory . $new_img_name;
+    
+                if (move_uploaded_file($tmpFileName, $img_upload_path)) {
+                    echo "File uploaded successfully. Stored at: $new_img_name";
+                } else {
+                    echo "Error uploading file.";
+                }
+    
+                dataInsertion($conn, $new_img_name, $studFname, $studMname, $studLname, $studLRN, $studSection, $studStrand, $studGender, $studAdviser, $studAge, $studToken, $studGrdLvl);
+    
+    
+                
             }
-
-            dataInsertion($conn, $new_img_name, $studFname, $studMname, $studLname,  $studSection, $studStrand, $studGender, $studAdviser, $studAge, $studToken, $studGrdLvl);
-
-
-            
+        }else{
+            echo "error".$_FILES['pfp']['error'];
         }
-    }else{
-        echo "error".$_FILES['pfp']['error'];
+    
+        }
+        
+ 
     }
 
-    }
-
-    function dataInsertion($conn, $pfp, $fname, $mname, $lname, $section, $strand, $gender, $adviser, $age, $token, $grdlvl)
+    function dataInsertion($conn, $pfp, $fname, $mname, $lname, $lrn, $section, $strand, $gender, $adviser, $age, $token, $grdlvl)
 {
-    $stmtInsertion = $conn->prepare("INSERT INTO students(`profile`, fname, minitial, lname, section, strand, gender, adviser, age, token, grd_lvl) VALUES(:pfp, :fname, :minitial, :lname, :section, :strand, :gender, :adviser, :age, :token, :grdlvl)");
+    $stmtInsertion = $conn->prepare("INSERT INTO students(`profile`, fname, minitial, lname, LRN, section, strand, gender, adviser, age, token, grd_lvl) VALUES(:pfp, :fname, :minitial, :lname, :LRN, :section, :strand, :gender, :adviser, :age, :token, :grdlvl)");
     $stmtInsertion->execute([
         ":pfp"=>$pfp,
         ":fname"=>$fname,
         ":minitial"=>$mname,
         ":lname"=>$lname,
+        ":LRN"=>$lrn,   
         ":section"=>$section,
         ":strand"=>$strand,
         ":gender"=>$gender,
@@ -67,6 +73,22 @@
 
 
 }
+
+    function userExist($conn, $fname, $mname, $lname) {
+        $stmtSelection = $conn->prepare("SELECT * FROM students WHERE fname = :fname AND minitial = :minitial AND lname = :lname");
+        $stmtSelection->execute([
+            ":fname"=>$fname,
+            ":minitial"=>$mname,
+            ":lname"=>$lname
+        ]);
+        $results = $stmtSelection->fetchAll(PDO::FETCH_ASSOC);
+        if($stmtSelection->rowCount() > 0) {
+            return true;
+        } else {
+            return false;
+        }
+        
+    }
 
 
 ?>
